+++ test/legacy/ut/unified_platform/resource/stream/ut_stream.cc
@@ -32,7 +32,7 @@ protected:
     {
         MOCKER(HrtGetStreamId).stubs().will(returnValue(fakeId));
         MOCKER(HrtGetDevice).stubs().will(returnValue(fakeDevLogId));
-        MOCKER(HrtGetDevicePhyIdByIndex).stubs().will(returnValue(fakeDevPhyId));
+        MOCKER(HrtGetDevicePhyIdByIndex).stubs().will(returnValue(static_cast<s32>(fakeDevPhyId)));
         MOCKER(HrtStreamCreateWithFlags).stubs().will(returnValue(fakePtr));
         MOCKER(HrtStreamGetSqId).stubs().will(returnValue(fakeSqId));
         MOCKER(HrtStreamDestroy).stubs();
